Tweaks to code, updates to changelog
