+++ src/bunit.web/Diffing/HtmlComparer.cs
@@ -12,7 +12,7 @@ using Bunit.Rendering;
 namespace Bunit.Diffing
 {
 	/// <summary>
-	/// Represents a test HTML comparer, that is configured to work with markup generated by the <see cref="TestRenderer"/> and <see cref="Htmlizer"/> classes.
+	/// Represents a test HTML comparer, that is configured to work with markup generated by the <see cref="ITestRenderer"/> and <see cref="Htmlizer"/> classes.
 	/// </summary>
 	public sealed class HtmlComparer
 	{
